requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
appdirs==1.4.4
astroid==2.11.7
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.6.0
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cssselect==1.1.0
decorator==5.1.1
dill==0.3.5.1
discord.py==1.7.3
fake-useragent==0.1.11
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
parse==1.19.0
pathspec==0.9.0
pip==22.1.2
platformdirs==2.5.2
pyee==8.2.2
pylint==2.14.4
pyppeteer==1.0.2
pyquery==1.4.3
python-dotenv==0.20.0
requests-html==0.10.0
requests==2.28.1
setuptools==63.1.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.10
validators==0.20.0
w3lib==1.22.0
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
